fix(scotus): Update scotus slip
Move slip code to chambers because it
still relies on the old method and i
didnt want to update it yet

Update scotus slip tests
